Plugin Directory: Avoid PHP Warnings when the uploaded zip has already been removed.
git-svn-id: https://meta.svn.wordpress.org/sites/trunk@13576 74240141-8908-4e6f-9713-ba540dce6ec7

+++ wordpress.org/public_html/wp-content/plugins/plugin-directory/jobs/class-zip-cleanup.php
@@ -88,9 +88,12 @@ class ZIP_Cleanup {
 			delete_post_meta( $plugin->ID, '_submitted_zip_loc' );
 
 			// Delete the file hash from the post.
-			$file_hash = sha1_file( get_attached_file( $attachment ) );
-			if ( $file_hash ) {
-				delete_post_meta( $plugin->ID, 'uploaded_zip_hash', $file_hash );
+			$file_on_disk = get_attached_file( $attachment );
+			if ( $file_on_disk && file_exists( $file_on_disk ) ) {
+				$file_hash = sha1_file( $file_on_disk );
+				if ( $file_hash ) {
+					delete_post_meta( $plugin->ID, 'uploaded_zip_hash', $file_hash );
+				}
 			}
 
 			// Include some log output for debugging.
